Develop

+++ src/controllers/MoviePrice.js
@@ -1,10 +1,15 @@
+/* eslint-disable no-useless-catch */
+// import Movie from '../model/Movie.js'
 import Movie from '../model/Movie.js'
 import MoviePrice from '../model/MoviePrice.js'
+import Seat from '../model/Seat.js'
 import { moviePriceService } from '../services/moviePrice.js'
+import Showtimes from '../model/Showtimes.js'
+
 import ApiError from '../utils/ApiError.js'
-import { slugify } from '../utils/stringToSlug.js'
+// import { slugify } from '../utils/stringToSlug.js'
 import {
-  moviePriceSchema,
+  // moviePriceSchema,
   updateMoviePriceSchema
 } from '../validations/MoviePrice.js'
 import { StatusCodes } from 'http-status-codes'
@@ -59,6 +64,27 @@ export const update = async (req, res, next) => {
     if (!id) {
       throw new ApiError(StatusCodes.NOT_FOUND, 'Id MoviePrice not found')
     }
+    //
+    const checkprice = await MoviePrice.findById(id)
+    const checkmovie = await Movie.find({ _id: checkprice.movieId })
+    // check xuat chieu
+    const checkshowtimes = await Showtimes.find({ movieId: checkprice.movieId })
+    const showtime = await checkshowtimes[0]
+    if (showtime != undefined) {
+      throw new ApiError(
+        StatusCodes.NOT_FOUND,
+        'Movies that are currently playing cannot be Update! (phim đang có xuất chiếu không thể sửa được )'
+      )
+    }
+    /// check status movie
+
+    const checkstt = checkmovie[0]
+    if (checkstt.status == 'COMING_SOON') {
+      throw new ApiError(
+        StatusCodes.NOT_FOUND,
+        'Phim đang công chiếu không thể sửa giá !'
+      )
+    }
     const { error } = updateMoviePriceSchema.validate(body, {
       abortEarly: true
     })
@@ -82,37 +108,8 @@ export const update = async (req, res, next) => {
 
 export const create = async (req, res, next) => {
   try {
-    const body = req.body
-    const { error } = moviePriceSchema.validate(body, { abortEarly: true })
-    if (error) {
-      throw new ApiError(StatusCodes.BAD_REQUEST, new Error(error).message)
-    }
+    const data = await moviePriceService.create(req.body)
 
-    // Check if a MoviePrice with the same movieId and dayType already exists
-    const existingMoviePrice = await MoviePrice.findOne({
-      movieId: body.movieId,
-      dayType: body.dayType
-    })
-    if (existingMoviePrice) {
-      throw new ApiError(
-        StatusCodes.BAD_REQUEST,
-        'A MoviePrice with the same movieId and dayType already exists'
-      )
-    }
-
-    const data = await MoviePrice.create({
-      ...body,
-      slug: slugify(body.name)
-    })
-    if (!data) {
-      throw new ApiError(StatusCodes.NOT_FOUND, 'Create MoviePrice failed')
-    }
-    // Update movie prices in movie collection
-    await Movie.findOneAndUpdate(data.movieId, {
-      $addToSet: {
-        prices: data._id
-      }
-    })
     return res.status(StatusCodes.CREATED).json({
       message: 'Success',
       data: data
@@ -125,16 +122,44 @@ export const create = async (req, res, next) => {
 export const remove = async (req, res, next) => {
   try {
     const id = req.params.id
+    const moviePrice = await moviePriceService.findById(id)
 
-    const data = await moviePriceService.remove(id)
+    if (!moviePrice) {
+      throw new ApiError(StatusCodes.NOT_FOUND, 'MoviePrice not found!')
+    }
+    const movie = await Movie.findById(moviePrice.movieId)
 
-    if (!data) {
-      throw new ApiError(StatusCodes.BAD_REQUEST, 'Delete MoviePrice failed!')
+    if (movie.status === 'IS_SHOWING') {
+      throw new ApiError(
+        StatusCodes.BAD_REQUEST,
+        'Movie is released. Cannot delete this price!'
+      )
     }
 
+    // Kiểm tra xem có chỗ nào được đặt trước cho bộ phim không
+    const seatBookings = await Seat.find({
+      movie: moviePrice.movieId,
+      seats: { $exists: true, $not: { $size: 0 } }
+    }).exec() // Giả sử việc đặt chỗ có mảng 'chỗ ngồi'
+    if (seatBookings.status === 'SOLD') {
+      throw new ApiError(
+        StatusCodes.BAD_REQUEST,
+        'Cannot delete price as there are seat bookings for this movie.'
+      )
+    }
+
+    await moviePriceService.remove(id)
+
+    // Cập nhật mảng giá trong Movie bằng cách loại bỏ giá đã xóa
+    const updatedMovie = await Movie.findByIdAndUpdate(
+      moviePrice.movie,
+      { $set: { prices: moviePrice._id } }, // Giả sử Movie có một mảng prices chứa các ID của MoviePrice
+      { new: [] } // Trả về Movie đã được cập nhật
+    )
+
     return res.status(StatusCodes.OK).json({
-      message: 'Success!',
-      data
+      message: 'MoviePrice deleted successfully!',
+      movie: updatedMovie
     })
   } catch (error) {
     next(error)
